+++ changelog.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [1.5.0]
 * Optimize reordering of scripts and styles via tree traversal.
+* Add Clutch widget support.
 
 ## [1.4.0]
 * Added "Include all dependencies to delayed scripts" feature.